Make sure grog works with single floating point (kx,ky) coordinates

+++ pygrappa/examples/basic_radialgrappaop.py
@@ -18,8 +18,8 @@ if __name__ == '__main__':
     # Radially sampled Shepp-Logan
     N, spokes, nc = 288, 72, 8
     kx, ky = radial(N, spokes)
-    kx = np.reshape(kx, (N, spokes), 'F').flatten()
-    ky = np.reshape(ky, (N, spokes), 'F').flatten()
+    kx = np.reshape(kx, (N, spokes), 'F').flatten().astype(float)
+    ky = np.reshape(ky, (N, spokes), 'F').flatten().astype(float)
     k = kspace_shepp_logan(kx, ky, ncoil=nc)
     k = whiten(k) # whitening seems to help conditioning of Gx, Gy
